move button and wrapper decoration to utils
