+++ express/code/blocks/ax-marquee/ax-marquee.js
@@ -1,11 +1,10 @@
-import { getLibs, toClassName, getIconElementDeprecated } from '../../scripts/utils.js';
+import { getLibs, toClassName, getIconElementDeprecated, addTempWrapperDeprecated, decorateButtonsDeprecated } from '../../scripts/utils.js';
 import BlockMediator from '../../scripts/block-mediator.min.js';
 import { addHeaderSizing } from '../../scripts/utils/location-utils.js';
 import {
   fetchPlanOnePlans,
   formatDynamicCartLink,
 } from '../../scripts/utils/pricing.js';
-import { addTempWrapperDeprecated, decorateButtonsDeprecated } from '../../scripts/utils/decorate.js';
 import { displayVideoModal, isVideoLink } from '../../scripts/widgets/video.js';
 
 let getMetadata; let createTag;
