+++ express/code/blocks/ax-table-of-contents/ax-table-of-contents.js
@@ -1,6 +1,5 @@
 /* eslint-disable import/named, import/extensions */
-import { getLibs, readBlockConfig } from '../../scripts/utils.js';
-import { addTempWrapperDeprecated } from '../../scripts/utils/decorate.js';
+import { getLibs, readBlockConfig, addTempWrapperDeprecated } from '../../scripts/utils.js';
 
 let createTag;
