+++ express/code/blocks/banner/banner.js
@@ -1,6 +1,6 @@
-import { getLibs, fixIcons } from '../../scripts/utils.js';
+import { getLibs, fixIcons, decorateButtonsDeprecated } from '../../scripts/utils.js';
 import { formatSalesPhoneNumber } from '../../scripts/utils/location-utils.js';
-import { decorateButtonsDeprecated, normalizeHeadings } from '../../scripts/utils/decorate.js';
+import { normalizeHeadings } from '../../scripts/utils/decorate.js';
 
 let createTag;
