+++ express/code/blocks/color-how-to-carousel/color-how-to-carousel.js
@@ -1,5 +1,4 @@
-import { getLibs, fixIcons } from '../../scripts/utils.js';
-import { addTempWrapperDeprecated } from '../../scripts/utils/decorate.js';
+import { getLibs, fixIcons, addTempWrapperDeprecated } from '../../scripts/utils.js';
 import isDarkOverlayReadable from '../../scripts/color-tools.js';
 
 let createTag;
