+++ express/code/blocks/icon-list/icon-list.js
@@ -1,5 +1,4 @@
-import { getLibs, toClassName, fixIcons, getIconElementDeprecated } from '../../scripts/utils.js';
-import { addTempWrapperDeprecated } from '../../scripts/utils/decorate.js';
+import { getLibs, toClassName, fixIcons, getIconElementDeprecated, addTempWrapperDeprecated } from '../../scripts/utils.js';
 
 let createTag;
