+++ express/code/blocks/pricing-cards-credits/pricing-cards-credits.js
@@ -1,5 +1,4 @@
-import { getLibs, fixIcons } from '../../scripts/utils.js';
-import { addTempWrapperDeprecated } from '../../scripts/utils/decorate.js';
+import { getLibs, fixIcons, addTempWrapperDeprecated } from '../../scripts/utils.js';
 
 let createTag;
