+++ express/code/blocks/pricing-cards/pricing-cards.js
@@ -3,8 +3,8 @@ import {
   yieldToMain,
   getIconElementDeprecated,
   fixIcons,
+  addTempWrapperDeprecated, decorateButtonsDeprecated,
 } from '../../scripts/utils.js';
-import { addTempWrapperDeprecated, decorateButtonsDeprecated } from '../../scripts/utils/decorate.js';
 import { debounce } from '../../scripts/utils/hofs.js';
 import { formatSalesPhoneNumber } from '../../scripts/utils/location-utils.js';
 import {
