+++ express/code/blocks/pricing-table/pricing-table.js
@@ -1,6 +1,6 @@
-import { getLibs, yieldToMain, fixIcons } from '../../scripts/utils.js';
+import { getLibs, yieldToMain, fixIcons, decorateButtonsDeprecated, addTempWrapperDeprecated } from '../../scripts/utils.js';
 import { debounce } from '../../scripts/utils/hofs.js';
-import { decorateButtonsDeprecated, addTempWrapperDeprecated, splitAndAddVariantsWithDash } from '../../scripts/utils/decorate.js';
+import { splitAndAddVariantsWithDash } from '../../scripts/utils/decorate.js';
 import { formatDynamicCartLink } from '../../scripts/utils/pricing.js';
 import { sendEventToAnalytics } from '../../scripts/instrument.js';
