+++ express/code/blocks/quick-action-hub/quick-action-hub.js
@@ -1,6 +1,5 @@
-import { getLibs, fixIcons } from '../../scripts/utils.js';
+import { getLibs, fixIcons, addTempWrapperDeprecated } from '../../scripts/utils.js';
 import { transformLinkToAnimation } from '../../scripts/utils/media.js';
-import { addTempWrapperDeprecated } from '../../scripts/utils/decorate.js';
 
 let createTag;
