+++ express/code/blocks/ratings/ratings.js
@@ -1,6 +1,6 @@
-import { getLibs, getLottie, lazyLoadLottiePlayer, toClassName, getIconElementDeprecated } from '../../scripts/utils.js';
+import { getLibs, getLottie, lazyLoadLottiePlayer, toClassName, getIconElementDeprecated, decorateButtonsDeprecated } from '../../scripts/utils.js';
 import BlockMediator from '../../scripts/block-mediator.min.js';
-import { decorateButtonsDeprecated, splitAndAddVariantsWithDash } from '../../scripts/utils/decorate.js';
+import { splitAndAddVariantsWithDash } from '../../scripts/utils/decorate.js';
 
 let createTag; let getConfig;
 let getMetadata;
