+++ express/code/blocks/toggle-bar/toggle-bar.js
@@ -1,5 +1,4 @@
-import { getLibs, fixIcons, readBlockConfig } from '../../scripts/utils.js';
-import { addTempWrapperDeprecated } from '../../scripts/utils/decorate.js';
+import { getLibs, fixIcons, readBlockConfig, addTempWrapperDeprecated } from '../../scripts/utils.js';
 import { sendEventToAnalytics, textToName } from '../../scripts/instrument.js';
 
 let createTag;
